+++ packages/framework/server/src/http/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @zh HTTP 模块导出
+ * @en HTTP module exports
+ */
+
+export * from './types.js';
+export { createHttpRouter } from './router.js';