+++ packages/framework/transaction/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @esengine/transaction
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies []:
+    - @esengine/server@4.1.0
+
 ## 2.0.5
 
 ### Patch Changes